+++ release_notes.md
@@ -1,6 +1,8 @@
 ## New Features
+- Support specification of max entity operation batch size
 
 ## Bug fixes
+- Fixed handling of function timeouts inside entity and activity functions and added tests
 - Skip constructor of AzureStorageDurabilityProvider if not used, to avoid spurious validation exceptions
 
 ## Breaking Changes